It builds, ship it
