+++ progress/impl/pom.xml
@@ -34,10 +34,11 @@
         </dependency>
         <dependency>
             <groupId>org.sakaiproject.kernel</groupId>
-            <artifactId>sakai-component-manager</artifactId>
+            <artifactId>sakai-kernel-api</artifactId>
+        </dependency>
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
+            <groupId>org.sakaiproject.kernel</groupId>
+            <artifactId>sakai-component-manager</artifactId>
         </dependency>
         <dependency>
             <groupId>org.sakaiproject.progress</groupId>
